Plots distribution of ages across survey years.

Plots distribution of ages across survey years.

age_frequency(
  data,
  subregions,
  common,
  sex = FALSE,
  cutoff = 0.95,
  facet_all = TRUE
)

Arguments

data

biological fisheries data containing age and sex information

subregions

choose NWFSC, PBS, AK Gulf, and/or AK BSAI

common

species common name

sex

TRUE or FALSE for if you want to differentiate by sex

cutoff

define a cutoff for grouping older ages together

facet_all

if TRUE this will facet all surveys regardless of missing data, if FALSE then only the region(s) with data will be faceted

Value

a ggplot object

Examples

if (FALSE) { # \dontrun{
data("nwfsc_bio")
data("afsc_bio")
data("pbs_bio")
all_data <- bind_rows(afsc_bio, nwfsc_bio, pbs_bio)

age_frequency(all_data, c("PBS", "NWFSC", "AK BSAI", "AK GULF"), "yellowtail rockfish")
} # }